addressing PR 10920 suggestions
